backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
